download composer

+++ project-baru/2CRUD-menu-pisah/index_author.php
@@ -4,7 +4,7 @@
 <head>
     <meta charset="UTF-8">
     <meta name="viewport" content="width=device-width, initial-scale=1.0">
-    <title>Index Book</title>
+    <title>Index Author</title>
 
 
     <link rel="stylesheet" href="https://stackpath.bootstrapcdn.com/bootstrap/4.5.2/css/bootstrap.min.css" integrity="sha384-JcKb8q3iqJ61gNV9KGb8thSsNjpSL0n8PARn9HuZOnIxN0hoP+VmmDGMN5t9UJ0Z" crossorigin="anonymous">
@@ -14,7 +14,7 @@
     <?php include "./author/query_author.php"; ?>
 
     <!-- TABLE -->
-    <a href="./author/add_author.php" class="btn btn-danger">Add New Author</a>
+    <a href="./author/add_author.php" class="btn btn-success">Add New Author</a>
 
     <div class="container">
         <table class="table">
@@ -39,7 +39,7 @@
                         <td><?php echo $row['address']; ?></td>
                         <td><?php echo $row['phone_number']; ?></td>
                         <td><a href="./author/edit_author.php?id=<?php echo $row['id']; ?>" class="btn btn-info">Edit</a>
-                            <a href="./author/query_author.php?delete=<?php echo $row['id']; ?>" class="btn btn-danger">Delete</a>
+                            <a href="./author/query_author.php?delete=<?php echo $row['id']; ?>" onclick="return confirm('Do You Want to Delete This?');" class="btn btn-danger">Delete</a>
                         </td>
                     </tr>
                 <?php endwhile; ?>
